% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{KCSKNNShiny}
\alias{KCSKNNShiny}
\title{K-Nearest Neighbour Classifier}
\usage{
KCSKNNShiny()
}
\description{
This is a KNN Classifier. It predicts any attribute(categorical) given a set of numeric input predictor values. Note that only numeric input predictors should be given.
             The k value can be chosen according to accuracies provided. The attribute to be predicted can be selected from the dropdown provided(select categorical attribute).
             This is because categorical attributes cannot be given as inputs here. A 'handsontable' is also provided to enter the input predictor values.
}
\examples{
if(interactive()){KCSKNNShiny()}

}
